# numguess-service/src/main/java/org/openapitools/api/GamesApi.java
/*
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.17.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package org.openapitools.api;

import org.openapitools.model.ApiRoot;
import org.openapitools.model.Error;
import org.openapitools.model.GameCreationResponse;
import org.openapitools.model.GamesCollection;
import org.springframework.lang.Nullable;
import java.net.URI;
import io.swagger.v3.oas.annotations.ExternalDocumentation;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.Parameters;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import io.swagger.v3.oas.annotations.media.ExampleObject;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.multipart.MultipartFile;

import jakarta.validation.Valid;
import jakarta.validation.constraints.*;
import java.util.List;
import java.util.Map;
import jakarta.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2025-11-10T20:40:10.675029-06:00[America/Chicago]", comments = "Generator version: 7.17.0")
@Validated
@Tag(name = "games", description = "Operations for managing game collections")
public interface GamesApi {

    default GamesApiDelegate getDelegate() {
        return new GamesApiDelegate() {};
    }

    String PATH_GAMES_GET = "/games";
    /**
     * GET /games : Get welcome page or games collection
     * Returns a welcome page (HTML) or games collection representation with hypermedia controls (JSON). For HTML clients, this displays a form to start a new game. For JSON clients, provides discoverable actions and navigation links (Level 3 HATEOAS). 
     *
     * @param accept Content type preference (optional, default to application/xhtml+xml)
     * @return Welcome page or games collection representation with hypermedia controls (status code 200)
     */
    @Operation(
        operationId = "gamesGet",
        summary = "Get welcome page or games collection",
        description = "Returns a welcome page (HTML) or games collection representation with hypermedia controls (JSON). For HTML clients, this displays a form to start a new game. For JSON clients, provides discoverable actions and navigation links (Level 3 HATEOAS). ",
        tags = { "games" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Welcome page or games collection representation with hypermedia controls", content = {
                @Content(mediaType = "application/xhtml+xml", schema = @Schema(implementation = String.class)),
                @Content(mediaType = "application/json", schema = @Schema(implementation = String.class)),
                @Content(mediaType = "application/hal+json", schema = @Schema(implementation = String.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = GamesApi.PATH_GAMES_GET,
        produces = { "application/xhtml+xml", "application/json", "application/hal+json" }
    )
    default ResponseEntity<String> gamesGet(
        @Parameter(name = "Accept", description = "Content type preference", in = ParameterIn.HEADER) @RequestHeader(value = "Accept", required = false, defaultValue = "application/xhtml+xml") String accept
    ) {
        return getDelegate().gamesGet(accept);
    }


    String PATH_GAMES_POST = "/games";
    /**
     * POST /games : Create a new game
     * Creates a new number guessing game instance and returns hypermedia controls for interaction. The game is initialized with a random number between 1 and 100. A UUID is generated to uniquely identify the new game instance.  Following Level 3 HATEOAS principles, the response includes comprehensive hypermedia links for all available actions: submitting guesses, deleting the game, or creating another game. 
     *
     * @param accept Content type preference (optional, default to application/xhtml+xml)
     * @return Game created successfully with hypermedia controls (status code 201)
     *         or Internal server error (status code 500)
     */
    @Operation(
        operationId = "gamesPost",
        summary = "Create a new game",
        description = "Creates a new number guessing game instance and returns hypermedia controls for interaction. The game is initialized with a random number between 1 and 100. A UUID is generated to uniquely identify the new game instance.  Following Level 3 HATEOAS principles, the response includes comprehensive hypermedia links for all available actions: submitting guesses, deleting the game, or creating another game. ",
        tags = { "games" },
        responses = {
            @ApiResponse(responseCode = "201", description = "Game created successfully with hypermedia controls", content = {
                @Content(mediaType = "application/xhtml+xml", schema = @Schema(implementation = String.class)),
                @Content(mediaType = "application/json", schema = @Schema(implementation = String.class)),
                @Content(mediaType = "application/hal+json", schema = @Schema(implementation = String.class))
            }),
            @ApiResponse(responseCode = "500", description = "Internal server error", content = {
                @Content(mediaType = "application/xhtml+xml", schema = @Schema(implementation = Error.class)),
                @Content(mediaType = "application/json", schema = @Schema(implementation = Error.class)),
                @Content(mediaType = "application/hal+json", schema = @Schema(implementation = Error.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = GamesApi.PATH_GAMES_POST,
        produces = { "application/xhtml+xml", "application/json", "application/hal+json" },
        consumes = { "application/x-www-form-urlencoded" }
    )
    default ResponseEntity<String> gamesPost(
        @Parameter(name = "Accept", description = "Content type preference", in = ParameterIn.HEADER) @RequestHeader(value = "Accept", required = false, defaultValue = "application/xhtml+xml") String accept
    ) {
        return getDelegate().gamesPost(accept);
    }


    String PATH_ROOT_GET = "/";
    /**
     * GET / : Get API root
     * Returns the API root resource with hypermedia controls (Level 3 HATEOAS). This endpoint serves as the entry point to the API, providing discoverable links to all top-level resources.  For HTML clients, may return HTTP 303 See Other redirect to &#x60;/games&#x60;. For JSON clients, returns hypermedia representation with navigational links. 
     *
     * @param accept Content type preference (optional, default to application/json)
     * @return API root representation with hypermedia controls (status code 200)
     *         or See Other - Redirect to /games (for HTML clients) (status code 303)
     */
    @Operation(
        operationId = "rootGet",
        summary = "Get API root",
        description = "Returns the API root resource with hypermedia controls (Level 3 HATEOAS). This endpoint serves as the entry point to the API, providing discoverable links to all top-level resources.  For HTML clients, may return HTTP 303 See Other redirect to `/games`. For JSON clients, returns hypermedia representation with navigational links. ",
        tags = { "games" },
        responses = {
            @ApiResponse(responseCode = "200", description = "API root representation with hypermedia controls", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ApiRoot.class)),
                @Content(mediaType = "application/hal+json", schema = @Schema(implementation = ApiRoot.class))
            }),
            @ApiResponse(responseCode = "303", description = "See Other - Redirect to /games (for HTML clients)")
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = GamesApi.PATH_ROOT_GET,
        produces = { "application/json", "application/hal+json" }
    )
    default ResponseEntity<ApiRoot> rootGet(
        @Parameter(name = "Accept", description = "Content type preference", in = ParameterIn.HEADER) @RequestHeader(value = "Accept", required = false, defaultValue = "application/json") String accept
    ) {
        return getDelegate().rootGet(accept);
    }

}